Remove new line indicator in the speaker and speech lists
For the speaker list this is identified with :/n and
all the items are the same so the code is straightfoward.

For the speech list, we have three types of items:
1) The first one is the first and last items on the list:
Since they were added last, they are not lists within the main list.

2) The rest of the items are lists within the main
list and most of these only contain one item within
that inner list, so we reference these by identifying
that we are calling only the first item.

3) Finally, part of the items that are lists within the main
list are a few other items that have more than one item within them.
These have to be dealt with differently since we need
to keep them as lists so we don't lose some of the speech to
other speakers when we create a dataframe