Create a new file for a shorter story
Using the full story consists of more than 300 unique
speakers and I'm afraid using the full story at the first go
may have hard-to-track errors.

Therefore, truncate the story such that it contains the first
5 appearing unique speakers with their respective speeches
as the dialogue happens and end before when the first time a sixth
speaker appears.