Copy and paste files from development folder
The development folder has folders such as a
virtual environment which make deployment of web page
difficult from github, therefore we opted to make
the development repo/folder private for now until we can
determine a good way to deploy without issues.

We use this repo as the main one so it's easy to
deploy the final results.